refactor: wrap a2a session logic in a class
