+++ src/handlers/ai/a2a/helpers.ts
@@ -2,7 +2,7 @@ import type { LanguageModelV2CallOptions } from '@ai-sdk/provider'
 
 import type { RemoteToolDefinition } from '../tools'
 
-import type { A2ASessionConfig } from './types'
+import type { A2ASessionConfig } from './config'
 
 export function safeResponseText(response: Response): Promise<string> {
   return response.text().catch(() => 'Unknown error')
