+++ src/test/factories/a2aSession.ts
@@ -1,28 +1,23 @@
 import { vi } from 'vitest'
 
-import type { ActiveA2ASession } from '@/handlers/ai/a2a/types'
+import { ActiveA2ASession } from '@/handlers/ai/a2a/session'
 
 export function createA2ASession(
   overrides?: Partial<ActiveA2ASession>
 ): ActiveA2ASession {
-  return {
-    reader: {
+  const session = new ActiveA2ASession(
+    {
       read: vi.fn(),
       cancel: vi.fn().mockResolvedValue(undefined),
       releaseLock: vi.fn(),
       closed: Promise.resolve(undefined),
     } as unknown as ReadableStreamDefaultReader<Uint8Array>,
-    contextId: undefined,
-    taskId: undefined,
-    sessionAbortController: new AbortController(),
-    config: { baseUrl: '', agentId: '', bearerToken: '' },
-    pendingToolRequests: new Map(),
-    unmatchedToolCalls: [],
-    unmatchedRemoteRequests: [],
-    sseBuffer: '',
-    readyToFinishForTools: false,
-    activeStreamArtifactId: undefined,
-    activeStreamContentType: undefined,
-    ...overrides,
+    undefined,
+    new AbortController(),
+    { baseUrl: '', agentId: '', bearerToken: '' }
+  )
+  if (overrides) {
+    Object.assign(session, overrides)
   }
+  return session
 }
